Signup Progress & Server Progress
! IMPORTANT !
We may have to convert all the files to .ejs instead of html. I will do this in a good way to fit the backend. For the front end guys please still use html for testing as it will work fine.

Added backend, now it's working with a server and mongodb. I will probally host a test version soon when I have the complete signup form working.