Updated NodeVersion & plugins from database
Updated the node modules for the backend to fit the current version of nodejs.

The plugins page now reads from our database. This is done in a safe way which means that a hacker won't be able to get into the users table.

I've been getting to know EJS language even more.