Add plugins to database
Fixed a login bugg, now they get it directly from the cookie without storing.

You can now add plugins to the database which comes up on the plugins page. Keep in mind that you need a database (mongodb) and nodejs in order to run the server and the database successfuly.