View Profile Completed
Completet view profile, you can now view other peoples profiles on their page.

Fixed some security issues with the passwords being sent to the client.

Added an extended input for the description when signing up. (Textarea)

If you view the users page everyone has now the same amount of space. (Fixed the description of the cards)

Commented some stuff to make it readable.

Changed the labels on the users page. (From "Contact {name}" to "See Profile". If someone press the see profile, then they will be taken to the users profile.

Fixed the links to the profiles on every plugin page. (For filters)

Minor bug fixes.

----

It's safe to say that we are soon getting to a beta state of the project.

// Olle.