Added "Join button" that sends email
If someone wants  to join the project. The host then gets an email sent to the set adress.

Added working join button on projects.

We can now send emails with our official email adress acc.

Added email.js file to handle all the email requests.

Email now gets passed on when logging in to the session, we can use this later in many situations.

When adding a plugin your email will get put into the database as the owner.